better defaults, cleaner utils (#140)
* refactor load_secrets to make it easier to manipulate secrets files

* NO_DOWNLOAD->DOWNLOAD, NO_MEMFS->AUTOSAVE with `no` being the default. LOGLEVEL defaults to DEBUG. ROOT_DIR defaults to . if not DOWNLOAD

* cleanup, incorporating some changes from #26 / #108

* include gitignore, symlink dockerignore 

* fix moved imports

* fix githook instruction 

* try adding a unit test?

* nicer ci step names
