+++ .github/Pipfile
@@ -19,6 +19,7 @@ prometheus_client = "*"
 black = "*"
 mypy = "*"
 pylint = "*"
+pytest = "*"
 protobuf = "*"
 types-termcolor = "*"
 types-protobuf = "*"
