+++ .github/workflows/checks.yml
@@ -18,10 +18,12 @@ jobs:
         uses: actions/cache@v1
         with:
           path: ~/.local/share/virtualenvs
-          key: '${{ runner.os }}-pipenv-v0-${{ hashFiles(''**/Pipfile.lock'') }}'
+          key: '${{ runner.os }}-pipenv-v1-${{ hashFiles(''**/Pipfile.lock'') }}'
       - name: Install dependencies
         if: steps.cache-pipenv.outputs.cache-hit != 'true'
         run: cp .github/Pipfile .; pipenv install --skip-lock --pre
+      - name: List dependency versions (check this if you get confusing results)
+        run: pipenv graph
       # - name: Setup tmate session # this lets you SSH into the container
       #   uses: mxschmitt/action-tmate@v3
       - name: Black
@@ -33,3 +35,5 @@ jobs:
       - name: Pylint
         run: pipenv run pylint --version; pipenv run pylint sample_bots mc_util forest mobfriend/mobfriend.py contact imogen/imogen.py
         if: always()
+      - name: Pytest
+        run: pipenv run python -m pytest
