+++ .gitignore
@@ -1,0 +1,4 @@
+*secrets
+*-cli
+data
+__pycache__
