+++ contact/teli.py
@@ -5,6 +5,7 @@
 import asyncio
 import logging
 import sys
+from asyncio.subprocess import PIPE, create_subprocess_exec
 from contextlib import asynccontextmanager
 from typing import Any, AsyncIterator, Optional
 
@@ -12,7 +13,7 @@ import aiohttp
 import phonenumbers as pn
 from aiohttp import web
 
-from forest.utils import get_secret, get_url
+from forest.utils import get_secret, APP_NAME
 
 
 def teli_format(raw_number: str) -> str:
@@ -149,6 +150,26 @@ class Teli:
         ) as resp:
             logging.info(resp)
             return await resp.json()
+
+
+@asynccontextmanager
+async def get_url(port: int = 8080) -> AsyncIterator[str]:
+    if not APP_NAME:
+        try:
+            logging.info("starting tunnel")
+            tunnel = await create_subprocess_exec(
+                *(f"lt -p {port}".split()),
+                stdout=PIPE,
+            )
+            assert tunnel.stdout
+            line = await tunnel.stdout.readline()
+            url = line.decode().lstrip("your url is: ").strip()
+            yield url + "/inbound"
+        finally:
+            logging.info("terminating tunnel")
+            tunnel.terminate()
+    else:
+        yield APP_NAME + ".fly.io"
 
 
 async def print_sms(raw_number: str, port: int = 8080) -> None:
